Group `torch.compile` utilities together
