+++ src/lightning/pytorch/core/module.py
@@ -34,7 +34,7 @@ from lightning.fabric.utilities.apply_func import convert_to_tensors
 from lightning.fabric.utilities.cloud_io import get_filesystem
 from lightning.fabric.utilities.device_dtype_mixin import _DeviceDtypeModuleMixin
 from lightning.fabric.utilities.distributed import _distributed_available, _sync_ddp
-from lightning.fabric.utilities.imports import _IS_WINDOWS, _TORCH_GREATER_EQUAL_2_0
+from lightning.fabric.utilities.imports import _IS_WINDOWS
 from lightning.fabric.wrappers import _FabricOptimizer
 from lightning.pytorch.callbacks.callback import Callback
 from lightning.pytorch.core.hooks import CheckpointHooks, DataHooks, ModelHooks
@@ -1639,86 +1639,6 @@ class LightningModule(
             self.__class__._register_load_state_dict_pre_hook(
                 weakref.proxy(self), pre_load_state_dict_hook, True  # type: ignore[arg-type]
             )
-
-    @classmethod
-    def from_compiled(cls, model: "torch._dynamo.OptimizedModule") -> "pl.LightningModule":
-        """Returns an instance LightningModule from the output of ``torch.compile``.
-
-        The ``torch.compile`` function returns a ``torch._dynamo.OptimizedModule``, which wraps the LightningModule
-        passed in as an argument, but doesn't inherit from it. This means that the output of ``torch.compile`` behaves
-        like a LightningModule but it doesn't inherit from it (i.e. `isinstance` will fail).
-
-        Use this method to obtain a LightningModule that still runs with all the optimizations from ``torch.compile``.
-        """
-        if not _TORCH_GREATER_EQUAL_2_0:
-            raise ModuleNotFoundError(f"`{cls.__name__}.from_compiled` requires torch>=2.0")
-
-        from torch._dynamo import OptimizedModule
-
-        if not isinstance(model, OptimizedModule):
-            raise ValueError(
-                f"`model` is required to be a `OptimizedModule`. Found a `{type(model).__name__}` instead."
-            )
-
-        orig_module = model._orig_mod
-
-        if not isinstance(orig_module, cls):
-            raise ValueError(
-                f"`model` is expected to be a compiled LightingModule. Found a `{type(orig_module).__name__}` instead"
-            )
-
-        orig_module._compiler_ctx = {
-            "compiler": "dynamo",
-            "dynamo_ctx": model.dynamo_ctx,
-            "original_forward": orig_module.forward,
-            "original_training_step": orig_module.training_step,
-            "original_validation_step": orig_module.validation_step,
-            "original_test_step": orig_module.test_step,
-            "original_predict_step": orig_module.predict_step,
-        }
-
-        orig_module.forward = model.dynamo_ctx(orig_module.forward)  # type: ignore[assignment]
-        orig_module.training_step = model.dynamo_ctx(orig_module.training_step)  # type: ignore[assignment]
-        orig_module.validation_step = model.dynamo_ctx(orig_module.validation_step)  # type: ignore[assignment]
-        orig_module.test_step = model.dynamo_ctx(orig_module.test_step)  # type: ignore[assignment]
-        orig_module.predict_step = model.dynamo_ctx(orig_module.predict_step)  # type: ignore[assignment]
-        return orig_module
-
-    @classmethod
-    def to_uncompiled(cls, model: Union["pl.LightningModule", "torch._dynamo.OptimizedModule"]) -> "pl.LightningModule":
-        """Returns an instance of LightningModule without any compilation optimizations from a compiled model.
-
-        This takes either a ``torch._dynamo.OptimizedModule`` returned by ``torch.compile()`` or a ``LightningModule``
-        returned by ``LightningModule.from_compiled``.
-
-        Note: this method will in-place modify the ``LightningModule`` that is passed in.
-        """
-        if not _TORCH_GREATER_EQUAL_2_0:
-            raise ModuleNotFoundError(f"`{cls.__name__}.to_uncompiled` requires torch>=2.0")
-
-        from torch._dynamo import OptimizedModule
-
-        if isinstance(model, OptimizedModule):
-            model = model._orig_mod
-
-        elif isinstance(model, cls):
-            if model._compiler_ctx is None:
-                raise ValueError(
-                    "`model` is required to be a compiled LightningModule. "
-                    "Found a non-compiled LightningModule instead."
-                )
-
-        else:
-            raise ValueError("`model` must either be an instance of OptimizedModule or LightningModule")
-
-        model.forward = model._compiler_ctx["original_forward"]
-        model.training_step = model._compiler_ctx["original_training_step"]
-        model.validation_step = model._compiler_ctx["original_validation_step"]
-        model.test_step = model._compiler_ctx["original_test_step"]
-        model.predict_step = model._compiler_ctx["original_predict_step"]
-        model._compiler_ctx = None
-
-        return model
 
 
 @contextmanager
