+++ src/lightning/pytorch/trainer/trainer.py
@@ -45,7 +45,6 @@ import lightning.pytorch as pl
 from lightning.fabric.utilities.apply_func import convert_tensors_to_scalars
 from lightning.fabric.utilities.cloud_io import get_filesystem
 from lightning.fabric.utilities.data import _auto_add_worker_init_fn
-from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_0
 from lightning.fabric.utilities.types import _PATH
 from lightning.fabric.utilities.warnings import PossibleUserWarning
 from lightning.pytorch.accelerators import Accelerator, TPUAccelerator
@@ -60,7 +59,7 @@ from lightning.pytorch.loops.fit_loop import _FitLoop
 from lightning.pytorch.loops.utilities import _parse_loop_limits, _reset_progress
 from lightning.pytorch.plugins import PLUGIN_INPUT, PrecisionPlugin
 from lightning.pytorch.profilers import Profiler
-from lightning.pytorch.strategies import DDPStrategy, FSDPStrategy, ParallelStrategy, SingleDeviceStrategy, Strategy
+from lightning.pytorch.strategies import ParallelStrategy, Strategy
 from lightning.pytorch.trainer import call, setup
 from lightning.pytorch.trainer.configuration_validator import verify_loop_configurations
 from lightning.pytorch.trainer.connectors.accelerator_connector import (
@@ -85,6 +84,7 @@ from lightning.pytorch.utilities.argparse import (
     parse_argparser,
     parse_env_variables,
 )
+from lightning.pytorch.utilities.compile import _maybe_unwrap_optimized, _verify_strategy_supports_compile
 from lightning.pytorch.utilities.data import has_len_all_ranks
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
 from lightning.pytorch.utilities.model_helpers import is_overridden
@@ -444,21 +444,6 @@ class Trainer:
         self._last_train_dl_reload_epoch = float("-inf")
         self._last_val_dl_reload_epoch = float("-inf")
 
-    def _maybe_unwrap_optimized(self, model: object) -> "pl.LightningModule":
-        if not _TORCH_GREATER_EQUAL_2_0:
-            if not isinstance(model, pl.LightningModule):
-                raise TypeError(f"`model` must be a `LightningModule`, got `{type(model).__qualname__}`")
-            return model
-        from torch._dynamo import OptimizedModule
-
-        if isinstance(model, OptimizedModule):
-            return model.from_compiled(model)
-        if isinstance(model, pl.LightningModule):
-            return model
-        raise TypeError(
-            f"`model` must be a `LightningModule` or `torch._dynamo.OptimizedModule`, got `{type(model).__qualname__}`"
-        )
-
     def fit(
         self,
         model: "pl.LightningModule",
@@ -485,7 +470,7 @@ class Trainer:
 
             datamodule: An instance of :class:`~lightning.pytorch.core.datamodule.LightningDataModule`.
         """
-        model = self._maybe_unwrap_optimized(model)
+        model = _maybe_unwrap_optimized(model)
         self.strategy._lightning_module = model
         call._call_and_handle_interrupt(
             self, self._fit_impl, model, train_dataloaders, val_dataloaders, datamodule, ckpt_path
@@ -571,7 +556,7 @@ class Trainer:
                     "`Trainer.validate()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._validate_impl, model, dataloaders, ckpt_path, verbose, datamodule
@@ -663,7 +648,7 @@ class Trainer:
                     "`Trainer.test()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._test_impl, model, dataloaders, ckpt_path, verbose, datamodule
@@ -757,7 +742,7 @@ class Trainer:
                     "`Trainer.predict()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._predict_impl, model, dataloaders, datamodule, return_predictions, ckpt_path
@@ -812,17 +797,7 @@ class Trainer:
     def _run(
         self, model: "pl.LightningModule", ckpt_path: Optional[_PATH] = None
     ) -> Optional[Union[_EVALUATE_OUTPUT, _PREDICT_OUTPUT]]:
-        if model._compiler_ctx is not None:
-            supported_strategies = [SingleDeviceStrategy, DDPStrategy, FSDPStrategy]
-            if self.strategy is not None and not any(isinstance(self.strategy, s) for s in supported_strategies):
-                supported_strategy_names = ", ".join(s.__name__ for s in supported_strategies)
-                raise RuntimeError(
-                    "Using a compiled model is incompatible with the current strategy: "
-                    f"{self.strategy.__class__.__name__}. "
-                    f"Only {supported_strategy_names} support compilation. "
-                    "Either switch to one of the supported strategies or avoid passing in "
-                    "a compiled model."
-                )
+        _verify_strategy_supports_compile(model, self.strategy)
 
         if self.state.fn == TrainerFn.FITTING:
             min_epochs, max_epochs = _parse_loop_limits(
